Add jwt user loader callback
